+++ src/frontend/query_protocol.ml
@@ -212,7 +212,7 @@ type _ t =
   | Extension_list : [ `All | `Enabled | `Disabled ] -> string list t
   | Path_list : [ `Build | `Source ] -> string list t
   | Occurrences (* *) :
-      [ `Ident_at of Msource.position ] * [ `Project | `Buffer ]
+      [ `Ident_at of Msource.position ] * [ `Project | `Buffer | `Renaming ]
       -> (Location.t list * occurrences_status) t
   | Signature_help : signature_help -> signature_help_result option t
       (** In current version, Merlin only uses the parameter [position] to answer