+++ src/ocaml-index/tests/tests-dirs/interfaces.t
@@ -22,3 +22,4 @@
    uid: Stdlib__Float.81; locs:
      "Float.t": File "main.mli", line 1, characters 9-16
    }, 0 approx shapes: {}, and shapes for CUS .
+  and related uids:{}